fix(mystic-ai): match the vendored ImGui's ImageButton signature
CI build failed: the bundled ImGui predates the str_id ImageButton overload. Use the (ImTextureID, size) form with PushID for a per-icon unique ID. Also cast the negative addon signature to uint32_t to silence C4245.

+++ modules/mystic-ai/entry.cpp
@@ -50,7 +50,7 @@ extern "C" __declspec(dllexport) AddonDefinition_t* GetAddonDef()
 {
     // Signature must be unique per addon — Mystic Clicker is -54321, Mystic
     // Trading is -54322.
-    AddonDef.Signature  = -54323;
+    AddonDef.Signature  = static_cast<uint32_t>(-54323);
     AddonDef.APIVersion = NEXUS_API_VERSION;
 
     AddonDef.Name             = "Mystic AI";
